requirements.txt
azure-mgmt==4.0.0